cli: add `ui.signed-off-by = true` to configure sign offs
Many of the desired features for the default commit template, notably
pre-populating things, can be solved with a simple pattern:

1. Inside `template-aliases.draft_commit_message`, use the following:

       if(config("name.space").as_boolean() && !description.contains("..."), ...)`

2. Set `name.space = true` in repos where you want it enabled.

This patch uses this idea to configure a new behavior to get
`Signed-off-by` lines included in their repositories.

While `Signed-off-by:` is a git mechanism, it is a common one and this
makes it easier to use jj OOTB with repositories that might require
those, without adding any more real surface area to the Rust codebase.
This general idea can be configured by the user though, e.g. for
`BUG=123` templates, or Gerrit `Change-Id`s, or other various things.

Signed-off-by: Austin Seipp <aseipp@pobox.com>